Create for_yerang_170925.py
This file is dedicated to Yerang, a student of the dait girls program in Sinsa, Seoul.
It changes old xls files to csv files with pandas manually with low efficiency.
But result mattered in this issue so I used this way.
I realized that when reading xls file in pure python, it changes into html form and 
can be parsed with bs4. That's how I did it.